+++ tests/test_violations/test_implementation.py
@@ -15,4 +15,4 @@ def test_visitor_returns_location():
 
 def test_checker_default_location():
     """Ensures that `BaseViolation` returns correct location."""
-    assert BaseViolation(None)._location() == (0, 0)
+    assert BaseViolation(None)._location() == (0, 0)  # noqa: Z441