+++ tests/test_visitors/test_ast/test_builtins/test_formatted_string.py
@@ -5,11 +5,14 @@ import pytest
 from wemake_python_styleguide.violations.consistency import (
     FormattedStringViolation,
 )
-from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
+from wemake_python_styleguide.visitors.ast.builtins import WrongStringVisitor
 
 regular_string = "'some value'"
+binary_string = "b'binary'"
+unicode_string = "u'unicode'"
 string_variable = "some = '123'"
 formated_string = "'x + y = {0}'.format(2)"
+procent_format = "'x = %d' % 1"
 key_formated_string = "'x + y = {res}'.format(res=2)"
 variable_format = """
 some = 'x = {0}'
@@ -22,8 +25,11 @@ f_empty_string = "f''"
 
 @pytest.mark.parametrize('code', [
     regular_string,
+    binary_string,
+    unicode_string,
     string_variable,
     formated_string,
+    procent_format,
     key_formated_string,
     variable_format,
 ])