+++ wemake_python_styleguide/visitors/presets/general.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
 from wemake_python_styleguide.visitors.ast import (
+    attributes,
+    builtins,
     comparisons,
     keywords,
     naming,
-    numbers,
 )
 from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
 from wemake_python_styleguide.visitors.ast.functions import (
@@ -14,7 +15,6 @@ from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
 from wemake_python_styleguide.visitors.ast.modules import (
     EmptyModuleContentsVisitor,
 )
-from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
 from wemake_python_styleguide.visitors.filenames.wrong_module_name import (
     WrongModuleNameVisitor,
 )
@@ -29,6 +29,8 @@ GENERAL_PRESET = (
     keywords.WrongTryFinallyVisitor,
     keywords.WrongExceptionTypeVisitor,
 
+    attributes.WrongAttributeVisitor,
+
     WrongFunctionCallVisitor,
     WrongImportVisitor,
 
@@ -36,8 +38,8 @@ GENERAL_PRESET = (
     naming.WrongModuleMetadataVisitor,
     naming.WrongVariableAssignmentVisitor,
 
-    numbers.MagicNumberVisitor,
-    WrongStringVisitor,
+    builtins.MagicNumberVisitor,
+    builtins.WrongStringVisitor,
 
     comparisons.WrongConditionalVisitor,
     comparisons.ComparisonSanityVisitor,